[MLIR] Move warp_execute_on_lane_0 from vector to gpu
